feat: added pagination on reviews, added properties - total review counts and total rating

+++ controllers/ReviewController.js
@@ -1,6 +1,46 @@
 const mongoose = require('mongoose')
 const { success, error, validation } = require('../common/responseAPI')
 const Review = require('../models/Review');
+const Product = require('../models/Product');
+
+
+
+// @desc    Get review
+// @route   GET /api/v1/review/
+// @access  Public
+
+exports.getReview = async (req, res) => {
+    try {
+        const { productId } = req.params;
+        const page = parseInt(req.query.page) || 1;
+        const limit = parseInt(req.query.limit) || 10;
+        const skip = (page - 1) * limit;
+
+        const reviews = await Review.find({ product: productId, isActive: true })
+            .skip(skip)
+            .limit(limit)
+            .exec();
+
+        const totalReviews = await Review.countDocuments({ product: productId, isActive: true });
+
+        const totalPages = Math.ceil(totalReviews / limit);
+
+        res.status(200).json(success("OK", {
+            reviews,
+            pagination: {
+                page_no: page,
+                per_page: limit,
+                total_reviews: totalReviews,
+                total_pages: totalPages,
+            },
+        },
+            res.statusCode),
+        );
+    } catch (err) {
+        return res.status(500).json(error("Something went wrong", res.statusCode));
+
+    }
+};
 
 
 // @desc    Create review
@@ -22,6 +62,9 @@ exports.addReview = async (req, res) => {
 
         const review = await Review.find({ product: productId, customer: customerId });
 
+        const product = await Product.findById({ _id: review[0].product });
+        if (!product) return res.status(404).json(error("Product not found", res.statusCode));
+
         let new_review = {}
 
         if (review.length === 0) { // add review if it does not exist else update the existing one
@@ -30,10 +73,18 @@ exports.addReview = async (req, res) => {
                 rating,
                 product: productId,
                 reviewer_name: `${req?.user?.firstName} ${req?.user?.lastName}`,
-                customer: customerId
+                customer: customerId,
             });
 
             new_review = await review.save();
+
+            let count = !product?.reviewCount ? 1 : Number(product?.reviewCount) + 1;
+            let totalRating = !product?.totalRating ? Number(rating) : (Number(product?.totalRating) + Number(rating)) / Number(count);
+
+            product.reviewCount = count;
+            product.totalRating = totalRating;
+            const updatedProduct = await product.save();
+
         } else {
             new_review = await Review.findByIdAndUpdate(
                 { _id: review[0]._id, customer: customerId },
@@ -68,8 +119,22 @@ exports.updateReview = async (req, res) => {
         if (!customerId) return res.status(400).json(error("Can't update review", res.statusCode));
 
         const review = await Review.find({ _id: req.params.id, customer: customerId });
+        console.log(review)
 
         if (!review) return res.status(404).json(error("Review not found", res.statusCode));
+
+        const product = await Product.findById({ _id: review[0].product });
+        if (!product) return res.status(404).json(error("Product not found", res.statusCode));
+        console.log(review)
+
+        if (review[0]?.rating && rating !== review[0].rating) {
+            // subtract contribution by previous rating and add contribution of new rating
+            let contributionOfPreviousRating = Number(review[0].rating) / Number(product.reviewCount);
+            let totalRating = !product?.totalRating ? Number(rating) : (Number(product.totalRating) - contributionOfPreviousRating) + (Number(rating) / Number(product.reviewCount));
+
+            product.totalRating = totalRating;
+            const updatedProduct = await product.save();
+        }
 
         const updatedReview = await Review.findByIdAndUpdate(
             { _id: req.params.id, customer: customerId },
@@ -87,18 +152,26 @@ exports.updateReview = async (req, res) => {
 };
 
 // @desc    Delete review
-// @route   PUT /api/v1/review/:id
+// @route   DELETE /api/v1/review/:id
 // @access  Private/Customer
 
 exports.deleteReview = async (req, res) => {
     try {
-        const { id } = req.params;
         const customerId = req.user._id;
         if (!customerId) return res.status(404).json(error("", res.statusCode));
 
         const review = await Review.find({ _id: req.params.id, customer: customerId });
-
         if (!review) return res.status(404).json(error("Review not found", res.statusCode));
+
+        const product = await Product.findById({ _id: review[0].product });
+        if (!product) return res.status(404).json(error("Product not found", res.statusCode));
+
+        let contributionOfPreviousRating = Number(review[0].rating) / Number(product.reviewCount);
+        let totalRating = (Number(product?.totalRating) - contributionOfPreviousRating);
+
+        product.reviewCount = Number(product.reviewCount) - 1;
+        product.totalRating = totalRating;
+        const updatedProduct = await product.save();
 
         let deletedReview = await Review.deleteOne({ _id: req.params.id });
